task_management/requirements.txt
Django>=4.2,<4.3
djangorestframework
psycopg2-binary
django-filter
djangorestframework-simplejwt
drf-yasg
python-decouple

